Macroize stream definitions
