--- src/stream/brotli/decoder.rs
@@ -1,73 +1,0 @@
-use std::{
-    io::Result,
-    pin::Pin,
-    task::{Context, Poll},
-};
-
-use bytes::Bytes;
-use futures::stream::Stream;
-use pin_project::unsafe_project;
-
-/// A brotli decoder, or decompressor.
-///
-/// This structure implements a [`Stream`] interface and will read compressed data from an
-/// underlying stream and emit a stream of uncompressed data.
-#[unsafe_project(Unpin)]
-#[derive(Debug)]
-#[cfg_attr(docsrs, doc(cfg(feature = "brotli")))]
-pub struct BrotliDecoder<S: Stream<Item = Result<Bytes>>> {
-    #[pin]
-    inner: crate::stream::generic::Decoder<S, crate::codec::BrotliDecoder>,
-}
-
-impl<S: Stream<Item = Result<Bytes>>> BrotliDecoder<S> {
-    /// Creates a new decoder which will read compressed data from the given stream and emit an
-    /// uncompressed stream.
-    pub fn new(stream: S) -> Self {
-        Self {
-            inner: crate::stream::generic::Decoder::new(stream, crate::codec::BrotliDecoder::new()),
-        }
-    }
-
-    /// Acquires a reference to the underlying stream that this decoder is wrapping.
-    pub fn get_ref(&self) -> &S {
-        self.inner.get_ref()
-    }
-
-    /// Acquires a mutable reference to the underlying stream that this decoder is wrapping.
-    ///
-    /// Note that care must be taken to avoid tampering with the state of the stream which may
-    /// otherwise confuse this decoder.
-    pub fn get_mut(&mut self) -> &mut S {
-        self.inner.get_mut()
-    }
-
-    /// Acquires a pinned mutable reference to the underlying stream that this decoder is wrapping.
-    ///
-    /// Note that care must be taken to avoid tampering with the state of the stream which may
-    /// otherwise confuse this decoder.
-    pub fn get_pin_mut<'a>(self: Pin<&'a mut Self>) -> Pin<&'a mut S> {
-        self.project().inner.get_pin_mut()
-    }
-
-    /// Consumes this decoder returning the underlying stream.
-    ///
-    /// Note that this may discard internal state of this decoder, so care should be taken
-    /// to avoid losing resources when this is called.
-    pub fn into_inner(self) -> S {
-        self.inner.into_inner()
-    }
-}
-
-impl<S: Stream<Item = Result<Bytes>>> Stream for BrotliDecoder<S> {
-    type Item = Result<Bytes>;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Result<Bytes>>> {
-        self.project().inner.poll_next(cx)
-    }
-}
-
-fn _assert() {
-    crate::util::_assert_send::<BrotliDecoder<Pin<Box<dyn Stream<Item = Result<Bytes>> + Send>>>>();
-    crate::util::_assert_sync::<BrotliDecoder<Pin<Box<dyn Stream<Item = Result<Bytes>> + Sync>>>>();
-}
