--- src/stream/brotli/mod.rs
@@ -1,4 +1,0 @@
-mod decoder;
-mod encoder;
-
-pub use self::{decoder::BrotliDecoder, encoder::BrotliEncoder};
