+++ src/stream/macros/mod.rs
@@ -1,0 +1,5 @@
+#[macro_use]
+mod encoder;
+
+#[macro_use]
+mod decoder;
