--- src/stream/zstd/mod.rs
@@ -1,4 +1,0 @@
-mod decoder;
-mod encoder;
-
-pub use self::{decoder::ZstdDecoder, encoder::ZstdEncoder};
